+++ src/js/IN_globalFilter.js
@@ -1,8 +1,8 @@
 var s = document.createElement('script');
-s.innerHTML = 'const LOADINGGIF = "' + chrome.extension.getURL('img/loading.svg') + '";';
+s.innerText = 'const LOADINGGIF = "' + chrome.extension.getURL('img/loading.svg') + '";';
 (document.head || document.documentElement).appendChild(s);
 s.onload = function() {
     s.parentNode.removeChild(s);
 };
 
-InjectMulti(['globalFilter.js', 'filters/default.js', 'filters/level.js', 'filters/craftable.js','LaunchFilter.js']);
+InjectMulti(['../lib/constructDOM.js' ,'globalFilter.js', 'filters/default.js', 'filters/level.js', 'filters/craftable.js','LaunchFilter.js']);